feat: move CardSkeleton
